refactor: add split logic file uploading logic to container
