+++ src/@modules/DependencyFormatter/containers/DependencyFormatter/DependencyFormatter.tsx
@@ -1,27 +1,15 @@
-import React, { useCallback } from 'react';
-import { Switch, Route, useRouteMatch, Redirect, useHistory } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
-import { FileSelector } from '../../components';
-import { recieveDependencies } from '../../actions';
+import React from 'react';
+import { Switch, Route, useRouteMatch, Redirect } from 'react-router-dom';
 import { DependencyRedactor } from '../DependencyRedactor';
+import { DependencySelector } from '../DependencySelector';
 
 export function DependencyFormatter() {
-  const { path, url } = useRouteMatch();
-  const history = useHistory();
-  const dispatch = useDispatch();
-
-  const uploadCallback = useCallback(
-    (nextDependencies) => {
-      dispatch(recieveDependencies(nextDependencies));
-      history.push(`${url}/format`);
-    },
-    [url, history, dispatch],
-  );
+  const { path } = useRouteMatch();
 
   return (
     <Switch>
       <Route path={`${path}/upload`}>
-        <FileSelector onFileLoaded={uploadCallback} />
+        <DependencySelector />
       </Route>
       <Route exact path={`${path}/format`}>
         <DependencyRedactor />
