+++ src/@modules/DependencyFormatter/containers/DependencySelector/index.ts
@@ -1,0 +1,1 @@
+export * from './DependencySelector';
